Backend/Server/requirements.txt
fastapi[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
passlib
pydantic
python-jose
bcrypt
